+++ src/dcase24t6/pre_processes/cnext.py
@@ -12,7 +12,12 @@ from torchoutil.nn.functional.get import get_device
 from dcase24t6.nn.ckpt import CNEXT_REGISTRY
 from dcase24t6.nn.encoders.convnext import convnext_tiny
 from dcase24t6.nn.functional import remove_index_nd
-from dcase24t6.pre_processes.common import batchify, is_audio_batch, unbatchify
+from dcase24t6.pre_processes.common import (
+    batchify,
+    is_audio_batch,
+    sanitize_batch,
+    unbatchify,
+)
 from dcase24t6.pre_processes.resample import Resample
 
 
@@ -29,6 +34,7 @@ class ResampleMeanCNext(nn.Module):
         offline: bool = False,
         device: str | torch.device | None = "cuda_if_available",
         input_time_dim: int = -1,
+        keep_batch: bool = False,
     ) -> None:
         device = get_device(device)
 
@@ -59,6 +65,7 @@ class ResampleMeanCNext(nn.Module):
         self.convnext = convnext
         self.resample = resample
         self.input_time_dim = input_time_dim
+        self.keep_batch = keep_batch
 
     @property
     def device(self) -> torch.device:
@@ -66,11 +73,15 @@ class ResampleMeanCNext(nn.Module):
 
     def forward(self, item_or_batch: dict[str, Any]) -> dict[str, Any]:
         if is_audio_batch(item_or_batch):
-            return self.forward_batch(item_or_batch)
+            batch = sanitize_batch(item_or_batch)
+            return self.forward_batch(batch)
+
+        item = item_or_batch
+        batch = batchify(item)
+        batch = self.forward_batch(batch)
+        if self.keep_batch:
+            return batch
         else:
-            item = item_or_batch
-            batch = batchify(item)
-            batch = self.forward_batch(batch)
             item = unbatchify(batch)
             return item