+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 ## [1.1.0] UNRELEASED
 ### Added
 - Download baseline weights during prepare script, and resume option in train and test scripts can now load baseline weights.
+- Hub for testing baseline model on a single file.
 
 ## [1.0.0] 2024-03-23
 ### Added